+++ src/main/java/br/edu/ufcg/preMatricula/controller/UsuarioController.java
@@ -15,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.edu.ufcg.preMatricula.exception.NaoAutorizadoException;
+import br.edu.ufcg.preMatricula.exception.RegisterNotFoundException;
 import br.edu.ufcg.preMatricula.model.RequestNote;
 import br.edu.ufcg.preMatricula.model.UsuarioNote;
 import br.edu.ufcg.preMatricula.service.UsuarioService;
@@ -24,6 +25,7 @@ import br.edu.ufcg.preMatricula.service.UsuarioService;
 @CrossOrigin(origins = "*")
 public class UsuarioController {
 	
+	private static final String COORDENADOR = "projsw@ccc.ufcg.edu.br";
 	@Autowired
 	UsuarioService usuarioService;
 	
@@ -48,27 +50,42 @@ public class UsuarioController {
 		RequestNote info = loginController.login(token);
 		if(info.getEmail() == null || !info.getHd().equals("ccc.ufcg.edu.br")){
 			throw new NaoAutorizadoException("Email Invalido!");
+		}else if(usuario.getEmail().equals(COORDENADOR)){
+			usuario.setAdmin(true);
+		}else if(usuario.isAdmin()){
+			usuario.setAdmin(false);
 		}
 		usuario.setIdToken(token);
 		usuarioService.save(usuario);
 		return usuario;
 	}
 	
-	@RequestMapping(value = "/usuario/{id}", method = RequestMethod.GET)
-	public ResponseEntity<UsuarioNote> getUsuario(@PathVariable("id") String id) {
-		UsuarioNote todo = usuarioService.getUsuarioId(id);
-		return new ResponseEntity<UsuarioNote>(todo, HttpStatus.OK);
+	@RequestMapping(value = "/usuario/login/{token}", method = RequestMethod.GET)
+	public ResponseEntity<UsuarioNote> usuarioLogin(@PathVariable("token") String token) throws UnsupportedOperationException, IOException {
+		UsuarioNote usuario = new UsuarioNote();
+		RequestNote info = loginController.login(token);
+		if(info.getEmail() == null ||!info.getHd().equals("ccc.ufcg.edu.br")){
+			throw new NaoAutorizadoException("Email Invalido!");
+		}
+		try{
+			usuario = usuarioService.getByToken(token);
+		}catch( RegisterNotFoundException e){
+			return new ResponseEntity<UsuarioNote>(usuario, HttpStatus.OK);
+		}
+		System.out.println(info.getFamily_name());
+		return new ResponseEntity<UsuarioNote>(usuario, HttpStatus.OK);
 	}
+	
 	
 	@RequestMapping(value = "/usuario/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<UsuarioNote> deleteUsuario(@PathVariable("id") String id) {
+	public ResponseEntity<UsuarioNote> deleteUsuario(@PathVariable("id") String id, @RequestHeader(required = true, value = LoginController.TOKEN_ID) String token) throws UnsupportedOperationException, IOException {
+		RequestNote login = loginController.login(token);
+		UsuarioNote usuario = usuarioService.getByToken(token);
+		if(!usuario.isAdmin()){
+			throw new NaoAutorizadoException("Operação Não atualizada");
+		}
 		UsuarioNote todo = usuarioService.delete(id);
 		return new ResponseEntity<UsuarioNote>(todo, HttpStatus.OK);
-	}
-	
-	@RequestMapping(value = "/usuario/email/{email}", method = RequestMethod.GET)
-	public UsuarioNote getDisciplinaCodigo(@PathVariable("email") String email) {
-		return usuarioService.getUsuario(email);
 	}
